frontend: Read app logo,title and app bar style from /config
We now send app config like logo,title and appBar style via /config
endpoint. Thus this patch allows to read from the endpoint and fallback
to default title and logo if nothing is returned from the endpoint.
Also To prevent harmful setting of innerHtml for svgs we should sanitize
the svg coming from backend. Thus this patch adds dompurify as a
dependency and updates the setting of innerHtml to be updated to use
sanitized html