backend: Read app title,logo file,header style from command line
As we want the flexibility to override the app logo,title based on
the config provided via command line flags. Thus This patch reads the
logo path,app title,header style from command line and also reads the
logo file which is than passed as string with the app title via /config
endpoint.